DEV-2262: move generation of urls to contribution admin view
